Fix conditionals (#686)
The Ruby on Rails guide contains the following example

```
  class Computer < ApplicationRecord
    validates :mouse, presence: true,
                      if: ["market.retail?", :desktop?],
                      unless: Proc.new { |c| c.trackpad.present? }
  end
```

Arrays and Strings were not supported before. https://github.com/rails/rails/blob/38fe5ae24476ef8808d5eb6366afe84ff43a3279/activesupport/lib/active_support/callbacks.rb#L429-L458
was used for inspiration.

Removes reference to "any object that responds to `#call`" in README, because
this is not how ActiveModel behaves. It doesn't check for `#call`, but
for class (`Proc`), and if it's an arbitrary object it calls callback
scope (which is `validate`). This behavior is not documented in Rails
guides.